requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# API Clients
slack-sdk==3.26.0
slack-bolt==1.18.0
notion-client==2.2.1
google-api-python-client==2.108.0
google-auth==2.25.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Vector Database
pinecone-client==3.0.0

# Embeddings & AI
openai==1.6.0
anthropic==0.7.0
tiktoken==0.5.2

# AWS Integration
boto3==1.34.0
botocore==1.34.0

# GCP Integration
google-cloud-dataflow==0.8.0
google-cloud-storage==2.14.0
apache-beam[gcp]==2.52.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# PII Detection & Security
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
cryptography==41.0.7

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
responses==0.24.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk==1.39.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
pre-commit==3.6.0
